fix add product in cart
