+++ ms-fe/src/components/admin/sale-couter/Bill.js
@@ -47,6 +47,7 @@ import FormUsingVoucher from "../../element/voucher/FormUsingVoucher";
 import numeral from "numeral";
 import SearchNameOrCodeVoucher from "../../element/voucher/SearchNameOrCodeVoucher";
 import { getToken } from "../../../service/Token";
+import TableOrderProduct from "./TableOrderProduct";
 
 const urlAutofillVoucher = "http://localhost:8080/api/client/autoFillVoucher";
 const baseUrl =
@@ -54,6 +55,7 @@ const baseUrl =
 
 const Bill = () => {
   var initialItems = [];
+  const [boolean, setBoolean] = useState(true);
   const [modalVisible, setModalVisible] = useState([]);
   const [modalAccountVisible, setModalAccountVisible] = useState([]);
   const [modalQRScanOpen, setModalQRScanOpen] = useState(false);
@@ -182,241 +184,241 @@ const Bill = () => {
     }
   };
 
-  const columns = [
-    {
-      title: "STT",
-      dataIndex: "index",
-      key: "stt",
-      width: 70,
-      render: (text, record, index) => {
-        return index + 1;
-      },
-    },
-    {
-      key: "product",
-      datatIndex: "product",
-      title: "Sản phẩm",
-      width: "50%",
-      render: (text, record, index) => {
-        return (
-          <Row style={{ width: "100%" }}>
-            <Col span={6} style={{ height: "100%" }}>
-              <div
-                style={{
-                  marginTop: "10px",
-                  marginRight: "10px",
-                }}
-              >
-                {record?.productDetail?.promotion?.length > 0 &&
-                record?.productDetail?.promotion[0]?.promotionValue ? (
-                  <Badge.Ribbon
-                    text={`Giảm ${
-                      record.productDetail.promotion[0].promotionValue
-                        ? record.productDetail.promotion[0].promotionMethod ===
-                          "%"
-                          ? record.productDetail.promotion[0].promotionValue +
-                            " " +
-                            record.productDetail.promotion[0].promotionMethod
-                          : record.productDetail.promotion[0].promotionValue.toLocaleString(
-                              "vi-VN",
-                              {
-                                style: "currency",
-                                currency: "VND",
-                              }
-                            )
-                        : null
-                    }`}
-                    color="red"
-                  >
-                    <Carousel style={{ maxWidth: "300px" }} autoplay>
-                      {record.productDetail.productImageResponse &&
-                        record.productDetail.productImageResponse.map(
-                          (item) => {
-                            return (
-                              <img
-                                key={item.id}
-                                style={{ width: "100%", marginTop: "10px" }}
-                                alt=""
-                                src={item.path}
-                              />
-                            );
-                          }
-                        )}
-                    </Carousel>
-                  </Badge.Ribbon>
-                ) : (
-                  <Carousel style={{ maxWidth: "300px" }} autoplay>
-                    {record?.productDetail?.productImageResponse &&
-                      record?.productDetail?.productImageResponse.map(
-                        (item) => {
-                          return (
-                            <img
-                              key={item.id}
-                              style={{ width: "100%", marginTop: "10px" }}
-                              alt=""
-                              src={item.path}
-                            />
-                          );
-                        }
-                      )}
-                  </Carousel>
-                )}
-              </div>
-            </Col>
-            <Col span={18} style={{ height: "100%" }}>
-              <div
-                className="m-5"
-                style={{
-                  textAlign: "start",
-                  height: "100%",
-                  justifyContent: "center",
-                }}
-              >
-                <span style={{ fontWeight: "500" }}>
-                  {record?.productDetail?.product?.productName +
-                    "-" +
-                    record?.productDetail?.brand?.brandName +
-                    "-" +
-                    record?.productDetail?.category?.categoryName +
-                    "-" +
-                    record?.productDetail?.button?.buttonName +
-                    "-" +
-                    record?.productDetail?.material?.materialName +
-                    "-" +
-                    record?.productDetail?.collar?.collarTypeName +
-                    "-" +
-                    record?.productDetail?.sleeve?.sleeveName +
-                    "-" +
-                    record?.productDetail?.shirtTail?.shirtTailTypeName +
-                    "-" +
-                    record?.productDetail?.pattern?.patternName +
-                    "-" +
-                    record?.productDetail?.form?.formName}
-                </span>
-                <br />
-                <div className={styles.optionColor}>
-                  <b>Màu sắc: </b>
-                  <span
-                    style={{
-                      backgroundColor: record?.productDetail?.color?.colorCode,
-                      marginLeft: "8px",
-                    }}
-                  ></span>
-                  {record?.productDetail?.color?.colorName}
-                </div>
-                <br />
-                <b>Kích cỡ: </b>
-                <span
-                  style={{
-                    marginLeft: "8px",
-                  }}
-                >
-                  {record?.productDetail?.size?.sizeName}
-                </span>
-              </div>
-            </Col>
-          </Row>
-        );
-      },
-    },
-    {
-      title: "Số lượng",
-      dataIndex: "quantity",
-      key: "quantity",
-      render: (text, record, index) => {
-        return (
-          <InputNumber
-            min={1}
-            max={record?.quantity >= record?.productDetail?.quantity}
-            value={record?.quantity}
-            onBlur={(event) =>
-              updateQuantity(record, index, event.target.value)
-            }
-          />
-        );
-      },
-    },
-    {
-      title: "Đơn giá",
-      dataIndex: "price",
-      key: "price",
+  // const columns = [
+  //   {
+  //     title: "STT",
+  //     dataIndex: "index",
+  //     key: "stt",
+  //     width: 70,
+  //     render: (text, record, index) => {
+  //       return index + 1;
+  //     },
+  //   },
+  //   {
+  //     key: "product",
+  //     datatIndex: "product",
+  //     title: "Sản phẩm",
+  //     width: "50%",
+  //     render: (text, record, index) => {
+  //       return (
+  //         <Row style={{ width: "100%" }}>
+  //           <Col span={6} style={{ height: "100%" }}>
+  //             <div
+  //               style={{
+  //                 marginTop: "10px",
+  //                 marginRight: "10px",
+  //               }}
+  //             >
+  //               {record?.productDetail?.promotion?.length > 0 &&
+  //               record?.productDetail?.promotion[0]?.promotionValue ? (
+  //                 <Badge.Ribbon
+  //                   text={`Giảm ${
+  //                     record.productDetail.promotion[0].promotionValue
+  //                       ? record.productDetail.promotion[0].promotionMethod ===
+  //                         "%"
+  //                         ? record.productDetail.promotion[0].promotionValue +
+  //                           " " +
+  //                           record.productDetail.promotion[0].promotionMethod
+  //                         : record.productDetail.promotion[0].promotionValue.toLocaleString(
+  //                             "vi-VN",
+  //                             {
+  //                               style: "currency",
+  //                               currency: "VND",
+  //                             }
+  //                           )
+  //                       : null
+  //                   }`}
+  //                   color="red"
+  //                 >
+  //                   <Carousel style={{ maxWidth: "300px" }} autoplay>
+  //                     {record.productDetail.productImageResponse &&
+  //                       record.productDetail.productImageResponse.map(
+  //                         (item) => {
+  //                           return (
+  //                             <img
+  //                               key={item.id}
+  //                               style={{ width: "100%", marginTop: "10px" }}
+  //                               alt=""
+  //                               src={item.path}
+  //                             />
+  //                           );
+  //                         }
+  //                       )}
+  //                   </Carousel>
+  //                 </Badge.Ribbon>
+  //               ) : (
+  //                 <Carousel style={{ maxWidth: "300px" }} autoplay>
+  //                   {record?.productDetail?.productImageResponse &&
+  //                     record?.productDetail?.productImageResponse.map(
+  //                       (item) => {
+  //                         return (
+  //                           <img
+  //                             key={item.id}
+  //                             style={{ width: "100%", marginTop: "10px" }}
+  //                             alt=""
+  //                             src={item.path}
+  //                           />
+  //                         );
+  //                       }
+  //                     )}
+  //                 </Carousel>
+  //               )}
+  //             </div>
+  //           </Col>
+  //           <Col span={18} style={{ height: "100%" }}>
+  //             <div
+  //               className="m-5"
+  //               style={{
+  //                 textAlign: "start",
+  //                 height: "100%",
+  //                 justifyContent: "center",
+  //               }}
+  //             >
+  //               <span style={{ fontWeight: "500" }}>
+  //                 {record?.productDetail?.product?.productName +
+  //                   "-" +
+  //                   record?.productDetail?.brand?.brandName +
+  //                   "-" +
+  //                   record?.productDetail?.category?.categoryName +
+  //                   "-" +
+  //                   record?.productDetail?.button?.buttonName +
+  //                   "-" +
+  //                   record?.productDetail?.material?.materialName +
+  //                   "-" +
+  //                   record?.productDetail?.collar?.collarTypeName +
+  //                   "-" +
+  //                   record?.productDetail?.sleeve?.sleeveName +
+  //                   "-" +
+  //                   record?.productDetail?.shirtTail?.shirtTailTypeName +
+  //                   "-" +
+  //                   record?.productDetail?.pattern?.patternName +
+  //                   "-" +
+  //                   record?.productDetail?.form?.formName}
+  //               </span>
+  //               <br />
+  //               <div className={styles.optionColor}>
+  //                 <b>Màu sắc: </b>
+  //                 <span
+  //                   style={{
+  //                     backgroundColor: record?.productDetail?.color?.colorCode,
+  //                     marginLeft: "8px",
+  //                   }}
+  //                 ></span>
+  //                 {record?.productDetail?.color?.colorName}
+  //               </div>
+  //               <br />
+  //               <b>Kích cỡ: </b>
+  //               <span
+  //                 style={{
+  //                   marginLeft: "8px",
+  //                 }}
+  //               >
+  //                 {record?.productDetail?.size?.sizeName}
+  //               </span>
+  //             </div>
+  //           </Col>
+  //         </Row>
+  //       );
+  //     },
+  //   },
+  //   {
+  //     title: "Số lượng",
+  //     dataIndex: "quantity",
+  //     key: "quantity",
+  //     render: (text, record, index) => {
+  //       return (
+  //         <InputNumber
+  //           min={1}
+  //           max={record?.quantity >= record?.productDetail?.quantity}
+  //           value={record?.quantity}
+  //           onBlur={(event) =>
+  //             updateQuantity(record, index, event.target.value)
+  //           }
+  //         />
+  //       );
+  //     },
+  //   },
+  //   {
+  //     title: "Đơn giá",
+  //     dataIndex: "price",
+  //     key: "price",
 
-      render: (text, record, index) => {
-        return (
-          <div style={{ textAlign: "center" }}>
-            {record?.productDetail?.promotionValue ? (
-              <span style={{ color: "#ccc" }}>
-                <strike>
-                  {record?.productDetail?.price?.toLocaleString("vi-VN", {
-                    style: "currency",
-                    currency: "VND",
-                  })}
-                </strike>
-              </span>
-            ) : (
-              <span>
-                {record?.productDetail?.price?.toLocaleString("vi-VN", {
-                  style: "currency",
-                  currency: "VND",
-                })}
-              </span>
-            )}
-            <br />
-            <span>
-              {record?.productDetail?.promotionValue
-                ? record?.productDetail?.promotionMethod === "%"
-                  ? (
-                      (record?.productDetail?.price *
-                        (100 - Number(record?.productDetail?.promotionValue))) /
-                      100
-                    )?.toLocaleString("vi-VN", {
-                      style: "currency",
-                      currency: "VND",
-                    })
-                  : (
-                      record?.productDetail?.price -
-                      Number(record?.productDetail?.promotionValue)
-                    )?.toLocaleString("vi-VN", {
-                      style: "currency",
-                      currency: "VND",
-                    })
-                : null}
-            </span>
-          </div>
-        );
-      },
-    },
-    {
-      title: "Thành tiền",
-      dataIndex: "totalPrice",
-      key: "totalPrice",
-      render: (text, record, index) => {
-        return (
-          <span>
-            {(record?.priceReduce * record?.quantity)?.toLocaleString("vi-VN", {
-              style: "currency",
-              currency: "VND",
-            })}
-          </span>
-        );
-      },
-    },
-    {
-      title: "Thao tác",
-      key: "action",
-      render: (text, record, index) => (
-        <Space size="middle">
-          <Button
-            icon={<DeleteOutlined />}
-            danger
-            href="#1"
-            key={record.key}
-            onClick={() => handleDeleteProduct(record, index)}
-          ></Button>
-        </Space>
-      ),
-    },
-  ];
+  //     render: (text, record, index) => {
+  //       return (
+  //         <div style={{ textAlign: "center" }}>
+  //           {record?.productDetail?.promotionValue ? (
+  //             <span style={{ color: "#ccc" }}>
+  //               <strike>
+  //                 {record?.productDetail?.price?.toLocaleString("vi-VN", {
+  //                   style: "currency",
+  //                   currency: "VND",
+  //                 })}
+  //               </strike>
+  //             </span>
+  //           ) : (
+  //             <span>
+  //               {record?.productDetail?.price?.toLocaleString("vi-VN", {
+  //                 style: "currency",
+  //                 currency: "VND",
+  //               })}
+  //             </span>
+  //           )}
+  //           <br />
+  //           <span>
+  //             {record?.productDetail?.promotionValue
+  //               ? record?.productDetail?.promotionMethod === "%"
+  //                 ? (
+  //                     (record?.productDetail?.price *
+  //                       (100 - Number(record?.productDetail?.promotionValue))) /
+  //                     100
+  //                   )?.toLocaleString("vi-VN", {
+  //                     style: "currency",
+  //                     currency: "VND",
+  //                   })
+  //                 : (
+  //                     record?.productDetail?.price -
+  //                     Number(record?.productDetail?.promotionValue)
+  //                   )?.toLocaleString("vi-VN", {
+  //                     style: "currency",
+  //                     currency: "VND",
+  //                   })
+  //               : null}
+  //           </span>
+  //         </div>
+  //       );
+  //     },
+  //   },
+  //   {
+  //     title: "Thành tiền",
+  //     dataIndex: "totalPrice",
+  //     key: "totalPrice",
+  //     render: (text, record, index) => {
+  //       return (
+  //         <span>
+  //           {(record?.priceReduce * record?.quantity)?.toLocaleString("vi-VN", {
+  //             style: "currency",
+  //             currency: "VND",
+  //           })}
+  //         </span>
+  //       );
+  //     },
+  //   },
+  //   {
+  //     title: "Thao tác",
+  //     key: "action",
+  //     render: (text, record, index) => (
+  //       <Space size="middle">
+  //         <Button
+  //           icon={<DeleteOutlined />}
+  //           danger
+  //           href="#1"
+  //           key={record.key}
+  //           onClick={() => handleDeleteProduct(record, index)}
+  //         ></Button>
+  //       </Space>
+  //     ),
+  //   },
+  // ];
 
   const options = [
     {
@@ -1669,6 +1671,8 @@ const Bill = () => {
                         cartId={cartId}
                         render={setRendered}
                         isEditProductTimeLine={false}
+                        setBoolean={setBoolean}
+                        boolean={boolean}
                       />
                     </Col>
                   </Row>
@@ -1676,7 +1680,7 @@ const Bill = () => {
                     className={styles.blackDivider}
                     style={{ marginTop: "3px" }}
                   />
-                  <Table
+                  {/* <Table
                     dataSource={
                       productDetails &&
                       productDetails?.map((record, index) => ({
@@ -1686,6 +1690,13 @@ const Bill = () => {
                     }
                     columns={columns}
                     pagination={false}
+                  /> */}
+                  <TableOrderProduct
+                    productDetails={productDetails}
+                    handleDeleteProduct={handleDeleteProduct}
+                    updateQuantity={updateQuantity}
+                    bool={boolean}
+                    cartId={cartId}
                   />
                 </div>
